requirements.txt
astroid==2.3.3
cachetools==3.1.1
certifi==2019.9.11
chardet==3.0.4
google-api-core==1.14.3
google-auth==1.7.0
google-cloud-vision==0.40.0
googleapis-common-protos==1.6.0
grpcio==1.25.0
idna==2.8
isort==4.3.21
lazy-object-proxy==1.4.3
mccabe==0.6.1
protobuf==3.10.0
pyasn1==0.4.7
pyasn1-modules==0.2.7
pylint==2.4.3
python-dotenv==0.10.3
pytz==2019.3
requests==2.22.0
rsa==4.0
six==1.13.0
urllib3==1.25.7
wrapt==1.11.2
